9-add_item.py

+++ 0x0B-python-input_output/9-add_item.py
@@ -4,25 +4,23 @@
 
 import json
 import sys
-load_from_json_file = __import__('8-load_from_json_file').load_from_json_file
-save_to_json_file = __import__('7-save_to_json_file').save_to_json_file
+import os
 
-def load_save:
-    """ this file is for comvert files to json"""
-    with open("add_item.json",'w+') as f:
-        count = 0
-        for line in f:
-            count += 1
-    if count == 0:
+save_to_json_file = __import__('7-save_to_json_file').save_to_json_file
+load_from_json_file = __import__('8-load_from_json_file').load_from_json_file
+""" this file is for comvert files to json"""
+if not os.path.isfile("add_item.json") :
+    with open('add_item.json', 'w+') as f:
         my_list = []
         for i in range(1, len(sys.argv)):
             my_list.append(sys.argv[i])
-            save_to_json_file(my_list,"add_item.json")
+        save_to_json_file(my_list, 'add_item.json')
+        print(my_list)
+else:
+    my_list = load_from_json_file('add_item.json')
+    for i in range(1, len(sys.argv)):
+        my_list.append(sys.argv[i])
+    save_to_json_file(my_list, 'add_item.json')
+    print(my_list)
+    
 
-    if len(load_from_json_file("add_item.json")) != 0:
-        my_list = load_from_json_file("add_item.json")
-        for i in range(1, len(sys.argv)):
-            my_list.append(sys.argv[i])
-    save_to_json_file(my_list,"add_item.json")
-if __name__ = '__main__':
-    def load_save
